Exclude `getInterpreter` from 'standalone.js'
